Standardize RestBuilder
- Use FgRestBuider in topic Authentication
- Write helper to transorm response errors to form errors
